refactor: centralize SVG to canvas logic
