+++ src/v2/utils/svgToCanvas.ts
@@ -1,0 +1,38 @@
+export async function renderSvgToCanvas(
+  svg: string,
+  width: number,
+  height: number,
+): Promise<HTMLCanvasElement> {
+  // フォントを読み込む
+  const { loadInterFonts } = await import('./fontLoader');
+  await loadInterFonts();
+
+  const blob = new Blob([svg], { type: 'image/svg+xml;charset=utf-8' });
+  const url = URL.createObjectURL(blob);
+
+  try {
+    const img = new Image();
+    img.crossOrigin = 'anonymous';
+    img.src = url;
+    await new Promise<void>((resolve, reject) => {
+      img.onload = () => resolve();
+      img.onerror = () => reject(new Error('Failed to load SVG'));
+    });
+
+    const canvas = document.createElement('canvas');
+    canvas.width = width * 2;
+    canvas.height = height * 2;
+    const ctx = canvas.getContext('2d');
+    if (!ctx) throw new Error('Failed to get canvas context');
+
+    // 背景を白で塗りつぶす
+    ctx.fillStyle = 'white';
+    ctx.fillRect(0, 0, canvas.width, canvas.height);
+    // SVG画像を描画
+    ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
+
+    return canvas;
+  } finally {
+    URL.revokeObjectURL(url);
+  }
+}
